Unit tests for CheckBucklescriptCompat.run()
