+++ src/__tests__/Esy_tests.bs.js
@@ -1,7 +1,7 @@
 // Generated by BUCKLESCRIPT, PLEASE EDIT WITH CARE
 'use strict';
 
-var Esy = require("../src/Esy.bs.js");
+var Esy = require("../Esy.bs.js");
 var Jest = require("@glennsl/bs-jest/src/jest.js");
 var Curry = require("bs-platform/lib/js/curry.js");
 var Filename = require("bs-platform/lib/js/filename.js");
